requirements.txt
Django==4.0.4
djangorestframework==3.13.1
pandas==1.3.4
requests==2.25.1
gunicorn==19.7.1
whitenoise==3.3.1
